requirements.txt
Django==4.2.0
djangorestframework==3.14.0
django-cors-headers==4.0.0
python-decouple==3.8
psycopg2-binary==2.9.6
Pillow==10.0.0
celery==5.3.0
redis==4.5.0
PyJWT==2.8.0
django-filter==23.2
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.2
django-model-utils==4.3.1
